modified:   tests/test.sh

+++ tests/test.sh
@@ -5,24 +5,27 @@ echo "Running tests..."
 echo
 make all
 echo  "exit" | ./myshell > out.txt &
+sleep 0.5
+
 if [ $? -eq 0 ] ; then
-  echo "Pass: Program exited zero\n"
+  echo "Pass: Program exited zero"
 else
-  echo "Fail: Program did not exit zero\n"
+  echo "Fail: Program did not exit zero"
   exit 1
 fi
 
-output=$(cat out.txt | tr -d '\n')
+actual=$(cat out.txt | tr -d '\n')
 pid=$(pgrep myshell)
-kill -9 $pid 
+if ! $pid; then 
+  kill -9 $pid  
+fi
 
-expected_output="$"
+expected="$"
 
-
-if [[ "$output" =~ "$expected_output" ]] ; then
-  echo "Pass: Output is correct"
+if [[ "$actual" =~ "$expected" ]] ; then
+  echo "Pass: output is correct"
 else
-  echo "Expected '$expected_output' but got: $output"
+  echo "Expected '$expected' but got: $actual"
   exit 1
 fi
